nice conflict handling

+++ test/test_helper.rb
@@ -15,20 +15,23 @@ module Kernel
   # Redirect standard out, standard error and the buffered logger for sprinkle to StringIO
   # capture_stdout { any_commands; you_want } => "all output from the commands"
   def capture_out
-    old_out, old_err = STDOUT.dup, STDERR.dup
-    stdout_read, stdout_write = IO.pipe
-    stderr_read, stderr_write = IO.pipe
-    $stdout.reopen(stdout_write)
-    $stderr.reopen(stderr_write)
-    yield
-    stdout_write.close
-    stderr_write.close
-    out = stdout_read.rewind && stdout_read.read rescue nil
-    err = stderr_read.rewind && stderr_read.read rescue nil
-    [out, err]
-  ensure
-    $stdout.reopen(old_out)
-    $stderr.reopen(old_err)
+    yield and return if ENV['DEBUG']
+    begin
+      old_out, old_err = STDOUT.dup, STDERR.dup
+      stdout_read, stdout_write = IO.pipe
+      stderr_read, stderr_write = IO.pipe
+      $stdout.reopen(stdout_write)
+      $stderr.reopen(stderr_write)
+      yield
+      stdout_write.close
+      stderr_write.close
+      out = stdout_read.rewind && stdout_read.read rescue nil
+      err = stderr_read.rewind && stderr_read.read rescue nil
+      [out, err]
+    ensure
+      $stdout.reopen(old_out)
+      $stderr.reopen(old_err)
+    end
   end
 end
 
@@ -43,8 +46,10 @@ class MiniTest::Spec
       "/tmp/gitdocs/#{c}"
     end
     pids = sub_paths.map { |path| fork do
-      STDOUT.reopen(File.open("/dev/null", 'w'))
-      STDERR.reopen(File.open("/dev/null", 'w'))
+      unless ENV['DEBUG']
+        STDOUT.reopen(File.open("/dev/null", 'w'))
+        STDERR.reopen(File.open("/dev/null", 'w'))
+      end
       begin
         Gitdocs::Runner.new(path, :growl => false, :polling_interval => 0.1).run
       rescue
@@ -60,6 +65,6 @@ class MiniTest::Spec
       pids.each { |pid| Process.kill("INT", pid) rescue nil }
     end
   ensure
-    FileUtils.rm_rf("/tmp/gitdocs")
+    FileUtils.rm_rf("/tmp/gitdocs") unless ENV['DEBUG']
   end
 end
